+++ src/Client/IconEffectWindow.cpp
@@ -1,0 +1,3 @@
+#include "IconEffectWindow.h"
+
+//SimplePlayer* IconEffects::getPlayer(IconType type, EffectUI* ui, bool second)